Refactor most i2c specifics to rpi file
